backend/requirements.txt
aiogram==3.3.0
python-dotenv==1.0.1
pydantic>=2.4.1,<2.6  # Диапазон, совместимый с aiogram 3.3.0